Installed react-router-dom. Created a components file. Added header, footer, and landing page components. Turned TodoList into a standalone component; removing all the related functional code from the App.js file while updating logic and imports. Added footer css rules, but haven't figured out how to apply them yet.
