+++ src/components/FooterComponent.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function FooterComponent() {
+  return (<div className="App-footer">Footer Component</div>);
+}
