+++ src/components/HeaderComponent.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function HeaderComponent() {
+  return (<div className='App-header'>Header Component</div>);
+}
