+++ src/components/LandingPage.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function LandingPage() {
+  return <div>Hi! welcome to my personal website.</div>;
+}
